# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orange Money API",
        "description": "API pour les services financiers mobile",
        "contact": {
            "email": "support@orangemoney.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ompay-4mgy.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/transactions/transfer": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un transfert d'argent",
                "operationId": "f5c88591ca5dacf77aaea5b9f6501116",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "receiver_phone",
                                    "amount"
                                ],
                                "properties": {
                                    "receiver_phone": {
                                        "type": "string",
                                        "example": "771234568"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Transfert familial"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert effectué avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/history": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Historique des transactions",
                "operationId": "8b99176af7f5e4dae053dfe23da0eed2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/compte/{id}/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Historique des transactions d'un compte",
                "operationId": "bbf2311e4824d423e9da7e5c99d67f16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du wallet/compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique des transactions du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transactions du compte"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "fees": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "reference": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/compte/{id}/payment-merchant": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un paiement à un marchand",
                "operationId": "0ed60ab81c11b8004cc6536ab679da87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du wallet/compte de l'émetteur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "merchant_identifier",
                                    "amount"
                                ],
                                "properties": {
                                    "merchant_identifier": {
                                        "type": "string",
                                        "example": "771234567"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Paiement marchand",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement marchand effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement effectué avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "fees": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "merchant_identifier": {
                                                    "type": "string"
                                                },
                                                "new_balance": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "merchant_new_balance": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "reference": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solde insuffisant"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Compte non trouvé ou marchand non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}